Technical/Update gem dependencies (#49)
* Updated runtime/development gem dependencies
* Updated reek/codeclimate configs
* Changed RSpec helpers namespaces
