+++ .circleci/gemspecs/latest
@@ -28,15 +28,15 @@ Gem::Specification.new do |spec|
 
   spec.add_development_dependency 'bundler-audit', '~> 0.9.1'
   spec.add_development_dependency 'fasterer', '~> 0.10.1'
-  spec.add_development_dependency 'ffaker', '~> 2.21'
+  spec.add_development_dependency 'ffaker', '~> 2.23'
   spec.add_development_dependency 'json_matchers', '~> 0.11.1'
   spec.add_development_dependency 'pry-byebug', '~> 3.10', '>= 3.10.1'
-  spec.add_development_dependency 'rake', '~> 13.0', '>= 13.0.6'
+  spec.add_development_dependency 'rake', '~> 13.1'
   spec.add_development_dependency 'reek', '~> 6.1', '>= 6.1.4'
   spec.add_development_dependency 'rspec', '~> 3.12'
-  spec.add_development_dependency 'rubocop', '~> 1.56'
-  spec.add_development_dependency 'rubocop-performance', '~> 1.19'
-  spec.add_development_dependency 'rubocop-rspec', '~> 2.23', '>= 2.23.2'
+  spec.add_development_dependency 'rubocop', '~> 1.57', '>= 1.57.2'
+  spec.add_development_dependency 'rubocop-performance', '~> 1.19', '>= 1.19.1'
+  spec.add_development_dependency 'rubocop-rspec', '~> 2.25'
   spec.add_development_dependency 'simplecov', '~> 0.22.0'
-  spec.add_development_dependency 'webmock', '~> 3.18', '>= 3.18.1'
+  spec.add_development_dependency 'webmock', '~> 3.19', '>= 3.19.1'
 end
