+++ truemail-client.gemspec
@@ -22,13 +22,16 @@ Gem::Specification.new do |spec|
     'bug_tracker_uri' => 'https://github.com/truemail-rb/truemail-ruby-client/issues'
   }
 
+  current_ruby_version = ::Gem::Version.new(::RUBY_VERSION)
+  ffaker_version = current_ruby_version >= ::Gem::Version.new('3.0.0') ? '~> 2.23' : '~> 2.21'
+
   spec.required_ruby_version = '>= 2.5.0'
-  spec.files = `git ls-files -z`.split("\x0").select { |f| f.match(%r{^(bin|lib)/|.ruby-version|truemail.gemspec|LICENSE}) }
+  spec.files = `git ls-files -z`.split("\x0").select { |f| f.match(%r{^(bin|lib)/|.ruby-version|truemail-client.gemspec|LICENSE}) }
   spec.require_paths = %w[lib]
 
-  spec.add_development_dependency 'ffaker', '~> 2.21'
+  spec.add_development_dependency 'ffaker', ffaker_version
   spec.add_development_dependency 'json_matchers', '~> 0.11.1'
-  spec.add_development_dependency 'rake', '~> 13.0', '>= 13.0.6'
+  spec.add_development_dependency 'rake', '~> 13.1'
   spec.add_development_dependency 'rspec', '~> 3.12'
-  spec.add_development_dependency 'webmock', '~> 3.18', '>= 3.18.1'
+  spec.add_development_dependency 'webmock', '~> 3.19', '>= 3.19.1'
 end
